requirements.txt
aioredis==1.3.1
altgraph==0.17
asgiref==3.5.2
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.6.1
Automat==20.2.0
cachetools==4.2.1
certifi==2022.6.15
cffi==1.15.1
channels==3.0.5
channels-redis==3.3.1
charset-normalizer==2.1.0
constantly==15.1.0
cryptography==37.0.4
daphne==3.0.2
Django==3.2.5
django-crispy-forms==1.13.0
djangorestframework==3.14.0
docopt==0.6.2
docutils==0.19
future==0.18.2
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
importlib-metadata==3.7.0
incremental==21.3.0
keyboard==0.13.5
Kivy==2.1.0
kivy-deps.angle==0.3.3
kivy-deps.glew==0.3.1
kivy-deps.sdl2==0.4.5
Kivy-examples==2.1.0
Kivy-Garden==0.1.5
MouseInfo==0.1.3
msgpack==1.0.4
numpy==1.21.6
opencv-python==4.7.0.68
pefile==2019.4.18
Pillow==8.2.0
pipreqs==0.4.11
psycopg2==2.9.3
py2exe==0.10.2.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.52
pycparser==2.21
pygame==2.0.1
PyGetWindow==0.0.9
Pygments==2.14.0
pyinstaller==4.2
pyinstaller-hooks-contrib==2020.11
PyMsgBox==1.0.9
pyOpenSSL==22.0.0
pyperclip==1.8.2
pypiwin32==223
PyRect==0.1.4
PyScreeze==0.1.27
PySimpleGUI==4.60.4
python-decouple==3.6
PyTweening==1.0.3
pytz==2021.1
pywin32==305
pywin32-ctypes==0.2.0
requests==2.28.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.1
Twisted==22.4.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing-extensions==3.7.4.3
urllib3==1.26.11
yarg==0.1.9
zipp==3.4.0
zope.interface==5.4.0
